requirements.txt
Django>=1.2
PyJWT>=1.1.0
cryptography>=0.9
lxml>=3.0.0
requests>=2.0.0